Updates app icons and launcher configuration
Updates the app icons for different platforms (macOS, web, Windows).

Configures the flutter_launcher_icons package to generate icons, disabling it for Android and iOS and explicitly setting icon sizes for web, Windows and macOS.

Also fixes a minor UI issue in the Edit Tasks page where the subject's text was not displayed.